backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.5.3
python-dotenv==1.0.1
httpx==0.26.0
structlog==24.1.0
pytest==7.4.3
pytest-asyncio==0.23.5
responses==0.24.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
groq==0.4.0  # For Groq API integration
